70% move of player and use skill

+++ Assets/Scripts/Skills/Skills.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Skills
 {
-    protected int factor;
+    protected float factor;
     protected int countDown;
     protected int countDownMax;
